adjust heading bar for mobile case
