+++ sky/dashboard/src/components/elements/layout.jsx
@@ -1,23 +1,9 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import { TopBar, SidebarProvider } from './sidebar';
+import { useMobile } from '@/hooks/useMobile';
 
 function LayoutContent({ children, highlighted }) {
-  const [isMobile, setIsMobile] = useState(false);
-
-  useEffect(() => {
-    const checkMobile = () => {
-      setIsMobile(window.innerWidth < 768);
-    };
-
-    // Check on initial load
-    checkMobile();
-
-    // Add resize listener
-    window.addEventListener('resize', checkMobile);
-
-    // Cleanup
-    return () => window.removeEventListener('resize', checkMobile);
-  }, []);
+  const isMobile = useMobile();
 
   return (
     <div className="min-h-screen bg-gray-50">
